Crate fedimint_dbdump

Macros

filter_prefixes ğŸ”’

push_db_key_items ğŸ”’

push_db_pair_items ğŸ”’

push_db_pair_items_no_serde ğŸ”’

try ğŸ”’

try ğŸ”’

Structs

Args ğŸ”’

DatabaseDump ğŸ”’

Structure to hold the deserialized structs from the database. Also includes metadata on which sections of the database to read.

SerdeWrapper ğŸ”’

__FieldVisitor ğŸ”’

__SerializeWith ğŸ”’

__Visitor ğŸ”’

Enums

__Field ğŸ”’

Constants

FIELDS ğŸ”’

RANGES ğŸ”’

USAGE ğŸ”’

Functions

main ğŸ”’

Opaque Types

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’